# Back_end/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "JQStock API",
    "description": "API para gestão de estoques, ruas, edificações e usuários",
    "version": "1.0.0",
    "contact": {
      "email": "gustavode.s.fonseca@gmail.com"
    }
  },
  "servers": [
    {
      "url": "https://wa-jqstock-api-dev.azurewebsites.net",
      "description": "Produção"
    },
    {
      "url": "http://localhost:8080",
      "description": "Desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/builds": {
      "get": {
        "tags": [
          "Prédios"
        ],
        "summary": "Listar edificações",
        "description": "Retorna lista paginada de edificações",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de edificações",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prédios"
        ],
        "summary": "Criar Prédio",
        "description": "Cadastra uma nova Prédio vinculada a uma rua específica",
        "parameters": [
          {
            "name": "StreetId",
            "in": "path",
            "description": "ID da rua onde a Prédio será criada",
            "required": true,
            "schema": {
              "type": "string",
              "example": "607f1f77bcf86cd799439012"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuildRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Edificação criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Build"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          }
        }
      }
    },
...
    "/floors": {
      "get": {
        "tags": [
          "Andares"
        ],
        "summary": "Listar andares",
        "description": "Retorna lista paginada de andares",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de andares",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Andares"
        ],
        "summary": "Criar andar",
        "description": "Cadastra um novo andar",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFloorRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Andar criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Floor"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Build": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439013"
          },
          "StreetId": {
            "type": "string",
            "example": "607f1f77bcf86cd799439012"
          },
          "name": {
            "type": "string",
            "example": "Building A"
          },
          "description": {
            "type": "string",
            "example": "Main building"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Za-z0-9]{3}$",
            "example": "BLD"
          },
          "Floors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de IDs dos andares desta Prédio"
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "Floor": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439013"
          },
          "BuildId": {
            "type": "string",
            "example": "607f1f77bcf86cd799439012"
          },
          "name": {
            "type": "string",
            "example": "Building A"
          },
          "description": {
            "type": "string",
            "example": "Main building"
          },
          "localCode": {
            "type": "string",
            "example": "A1"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Za-z0-9]{3}$",
            "example": "BLD"
          },
          "stockedItems": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de IDs dos itens estocados neste andar"
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
...
      "CreateBuildRequest": {
        "type": "object",
        "required": [
          "name",
          "code",
          "createdBy"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Nova Edificação"
          },
          "description": {
            "type": "string",
            "example": "Edificação principal"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Za-z0-9]{3}$",
            "example": "BLD"
          },
          "createdBy": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          }
        }
      },
...
      "BuildListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Build"
            }
          },
          "totalItems": {
            "type": "integer",
            "example": 15
          },
          "totalPages": {
            "type": "integer",
            "example": 2
          },
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
      "FloorListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Floor"
            }
          },
          "totalItems": {
            "type": "integer",
            "example": 15
          },
          "totalPages": {
            "type": "integer",
            "example": 2
          },
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "CreateFloorRequest": {
        "type": "object",
        "required": [
          "BuildId",
          "name",
          "code",
          "createdBy"
        ],
        "properties": {
          "BuildId": {
            "type": "string",
            "example": "607f1f77bcf86cd799439013"
          },
          "name": {
            "type": "string",
            "example": "New Floor"
          },
          "description": {
            "type": "string",
            "example": "Additional floor"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Za-z0-9]{3}$",
            "example": "FL2"
          },
          "createdBy": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Mensagem de erro"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "code"
                },
                "message": {
                  "type": "string",
                  "example": "Código inválido"
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "description": "Número da página",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1,
          "example": 1
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "Itens por página",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 10,
          "example": 10
        }
      },
...
    },
    "responses": {
      "BadRequestError": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "InvalidCode": {
                "value": {
                  "message": "Código inválido - deve ter 3 caracteres alfanuméricos",
                  "status": 400
                }
              }
            }
          }
        }
      },
...
    }
  }
}